requirements.txt
smartapi-python
pyotp
requests
feedparser
pandas